refactor(requests): collapse 4 create routes into canonical /requests/new (closes #1518)
- /requests/new: now canonical (full form, was a redirect before)
- /requests/create: redirect to /requests/new (was the form)
- /(tabs)/create: redirect to /requests/new (already was)
- updated 3 navigation refs: app/index.tsx, InlineOtpFlow returnTo default,
  AppShell comment
- tab-bar entry stays (href: null) so route exists for deep links
