+++ app/requests/create.tsx
@@ -1,487 +1,10 @@
-import { useState, useEffect, useCallback } from "react";
-import {
-  View,
-  Text,
-  ScrollView,
-  ActivityIndicator,
-  Alert,
-  Platform,
-  Pressable,
-} from "react-native";
-import { SafeAreaView } from "react-native-safe-area-context";
-import { useRouter, useLocalSearchParams } from "expo-router";
-import { useTypedRouter, ROUTES } from "@/lib/navigation";
-import { MapPin, ChevronLeft, Inbox } from "lucide-react-native";
-import Button from "@/components/ui/Button";
-import Input from "@/components/ui/Input";
-import EmptyState from "@/components/ui/EmptyState";
-import { api, apiPost } from "@/lib/api";
-import { useAuth } from "@/contexts/AuthContext";
-import { colors } from "@/lib/theme";
-import CityFnsServicePicker, {
-  CityOption,
-  FnsOption,
-  ServiceOption,
-} from "@/components/requests/CityFnsServicePicker";
-import InlineOtpFlow from "@/components/requests/InlineOtpFlow";
-import { draftStorage } from "@/lib/draftStorage";
+import { Redirect } from "expo-router";
 
-// Single canonical key (v1). Replaces legacy "pending_request_draft".
-const DRAFT_KEY = "p2ptax_request_draft_v1";
-const LEGACY_DRAFT_KEY = "pending_request_draft";
-
-interface RequestDraft {
-  title: string;
-  description: string;
-  cityId: string | null;
-  citySlug: string | null;
-  fnsId: string | null;
-  serviceId: string | null;
-}
-
-export default function CreateRequest() {
-  const router = useRouter();
-  const nav = useTypedRouter();
-  const { isAuthenticated, isLoading: authLoading } = useAuth();
-  const params = useLocalSearchParams<{ restore?: string }>();
-  const restoreMode = params.restore === "1";
-
-  const [title, setTitle] = useState("");
-  const [description, setDescription] = useState("");
-  const [selectedCityId, setSelectedCityId] = useState<string | null>(null);
-  const [selectedFnsId, setSelectedFnsId] = useState<string | null>(null);
-  const [selectedServiceId, setSelectedServiceId] = useState<string | null>(null);
-
-  const [cities, setCities] = useState<CityOption[]>([]);
-  const [fnsOffices, setFnsOffices] = useState<FnsOption[]>([]);
-  const [services, setServices] = useState<ServiceOption[]>([]);
-
-  const [cityOpen, setCityOpen] = useState(false);
-  const [fnsOpen, setFnsOpen] = useState(false);
-  const [serviceOpen, setServiceOpen] = useState(false);
-  const [loadingFns, setLoadingFns] = useState(false);
-  const [submitting, setSubmitting] = useState(false);
-  const [loadingInit, setLoadingInit] = useState(true);
-  const [loadError, setLoadError] = useState(false);
-  const [submitted, setSubmitted] = useState(false);
-  const [submitError, setSubmitError] = useState("");
-  const [atLimit, setAtLimit] = useState(false);
-  const [draftLoaded, setDraftLoaded] = useState(false);
-  const [showOtpFlow, setShowOtpFlow] = useState(false);
-
-  // Load cities/services — public, no auth required.
-  useEffect(() => {
-    async function init() {
-      try {
-        const [citiesRes, servicesRes] = await Promise.all([
-          api<{ items: CityOption[] }>("/api/cities", { noAuth: true }),
-          api<{ items: ServiceOption[] }>("/api/services", { noAuth: true }),
-        ]);
-        setCities(citiesRes.items);
-        setServices(servicesRes.items);
-      } catch {
-        setLoadError(true);
-      } finally {
-        setLoadingInit(false);
-      }
-    }
-    init();
-  }, []);
-
-  // Check request limit when authenticated.
-  useEffect(() => {
-    if (!isAuthenticated || authLoading) return;
-    api<{ requestsUsed: number; requestsLimit: number }>("/api/dashboard/stats")
-      .then((stats) => {
-        if (stats.requestsUsed >= stats.requestsLimit) {
-          setAtLimit(true);
-        }
-      })
-      .catch(() => {});
-  }, [isAuthenticated, authLoading]);
-
-  const loadFnsForCity = useCallback(async (citySlug: string) => {
-    setLoadingFns(true);
-    setFnsOffices([]);
-    try {
-      const data = await api<{
-        city: { id: string; name: string; slug: string };
-        items: FnsOption[];
-      }>(`/api/cities/${citySlug}/ifns`, { noAuth: true });
-      setFnsOffices(data.items);
-    } catch {
-      /* silent */
-    } finally {
-      setLoadingFns(false);
-    }
-  }, []);
-
-  // Restore draft on mount (anyone — anon or returning post-login).
-  // Reads new key first, falls back to legacy key for in-flight users.
-  useEffect(() => {
-    if (loadingInit || draftLoaded) return;
-    (async () => {
-      try {
-        let raw = await draftStorage.get(DRAFT_KEY);
-        if (!raw) {
-          raw = await draftStorage.get(LEGACY_DRAFT_KEY);
-          if (raw) {
-            // Migrate legacy → v1 key, then drop the old one.
-            await draftStorage.set(DRAFT_KEY, raw);
-            await draftStorage.del(LEGACY_DRAFT_KEY);
-          }
-        }
-        if (raw) {
-          const draft: RequestDraft = JSON.parse(raw);
-          if (draft.title) setTitle(draft.title);
-          if (draft.description) setDescription(draft.description);
-          if (draft.cityId) setSelectedCityId(draft.cityId);
-          if (draft.fnsId) setSelectedFnsId(draft.fnsId);
-          if (draft.serviceId) setSelectedServiceId(draft.serviceId);
-          if (draft.citySlug) loadFnsForCity(draft.citySlug);
-        }
-      } catch {
-        /* ignore parse / storage errors */
-      } finally {
-        setDraftLoaded(true);
-        if (restoreMode) {
-          // legacy returnTo path — already restored, nothing else to do.
-        }
-      }
-    })();
-  }, [loadingInit, draftLoaded, restoreMode, loadFnsForCity]);
-
-  // Persist draft on every form-field change (after initial load to avoid wiping).
-  useEffect(() => {
-    if (!draftLoaded) return;
-    const selectedCity = cities.find((c) => c.id === selectedCityId);
-    const draft: RequestDraft = {
-      title,
-      description,
-      cityId: selectedCityId,
-      citySlug: selectedCity?.slug ?? null,
-      fnsId: selectedFnsId,
-      serviceId: selectedServiceId,
-    };
-    const isEmpty =
-      !draft.title &&
-      !draft.description &&
-      !draft.cityId &&
-      !draft.fnsId &&
-      !draft.serviceId;
-    if (isEmpty) return;
-    draftStorage.set(DRAFT_KEY, JSON.stringify(draft)).catch(() => {});
-  }, [
-    draftLoaded,
-    title,
-    description,
-    selectedCityId,
-    selectedFnsId,
-    selectedServiceId,
-    cities,
-  ]);
-
-  const handleCitySelect = useCallback(
-    (city: CityOption) => {
-      setSelectedCityId(city.id);
-      setSelectedFnsId(null);
-      setCityOpen(false);
-      loadFnsForCity(city.slug);
-    },
-    [loadFnsForCity]
-  );
-
-  const handleFnsSelect = useCallback((fns: FnsOption) => {
-    setSelectedFnsId(fns.id);
-    setFnsOpen(false);
-  }, []);
-
-  const handleServiceSelect = useCallback((svc: ServiceOption) => {
-    setSelectedServiceId(svc.id);
-    setServiceOpen(false);
-  }, []);
-
-  const titleValid = title.trim().length >= 3 && title.trim().length <= 100;
-  const descriptionValid =
-    description.trim().length >= 10 && description.trim().length <= 2000;
-
-  const formValid =
-    titleValid && descriptionValid && !!selectedCityId && !!selectedFnsId && !atLimit;
-
-  // Actually post the request — used by both auth and post-OTP paths.
-  const submitRequestAuthed = useCallback(async () => {
-    setSubmitting(true);
-    setSubmitError("");
-    try {
-      const result = await apiPost<{ id: string }>("/api/requests", {
-        title: title.trim(),
-        cityId: selectedCityId,
-        fnsId: selectedFnsId,
-        serviceId: selectedServiceId || undefined,
-        description: description.trim(),
-        files: [],
-      });
-      // Clear draft on success.
-      await draftStorage.del(DRAFT_KEY).catch(() => {});
-      await draftStorage.del(LEGACY_DRAFT_KEY).catch(() => {});
-
-      const goToDetail = () => nav.replaceAny(`/requests/${result.id}/detail`);
-      if (Platform.OS === "web") {
-        goToDetail();
-      } else {
-        Alert.alert(
-          "Заявка опубликована",
-          "Специалисты по вашей ФНС увидят её и напишут вам. Обычно первый отклик приходит в течение 24 часов.",
-          [{ text: "OK", onPress: goToDetail }],
-          { onDismiss: goToDetail }
-        );
-      }
-    } catch (e: unknown) {
-      const msg =
-        e instanceof Error
-          ? e.message
-          : "Не удалось опубликовать заявку. Проверьте данные и попробуйте ещё раз.";
-      setSubmitError(msg);
-    } finally {
-      setSubmitting(false);
-    }
-  }, [title, description, selectedCityId, selectedFnsId, selectedServiceId, nav]);
-
-  const handleSubmit = useCallback(async () => {
-    setSubmitted(true);
-    setSubmitError("");
-    if (!formValid || submitting) return;
-
-    if (isAuthenticated) {
-      await submitRequestAuthed();
-      return;
-    }
-
-    // Anonymous path — open inline OTP block.
-    setShowOtpFlow(true);
-  }, [formValid, submitting, isAuthenticated, submitRequestAuthed]);
-
-  const selectedCity = cities.find((c) => c.id === selectedCityId);
-  const selectedFns = fnsOffices.find((f) => f.id === selectedFnsId);
-  const selectedService = services.find((s) => s.id === selectedServiceId);
-
-  if (authLoading || loadingInit) {
-    return (
-      <SafeAreaView className="flex-1 bg-surface2">
-        <View className="flex-1 items-center justify-center">
-          <ActivityIndicator size="large" color={colors.primary} />
-        </View>
-      </SafeAreaView>
-    );
-  }
-
-  if (loadError && cities.length === 0) {
-    return (
-      <SafeAreaView className="flex-1 bg-surface2">
-        <View className="flex-1 items-center justify-center">
-          <EmptyState
-            icon={MapPin}
-            title="Не удалось загрузить данные"
-            subtitle="Проверьте соединение и попробуйте снова"
-          />
-        </View>
-      </SafeAreaView>
-    );
-  }
-
-  // Authenticated user at the active-request limit (5/5). Show EmptyState
-  // only — no form. Anonymous users still see the form (they haven't hit
-  // any limit yet). Wave 6 R3.
-  if (isAuthenticated && atLimit) {
-    return (
-      <SafeAreaView className="flex-1 bg-surface2">
-        <View className="px-4 pt-4">
-          <Pressable
-            accessibilityRole="button"
-            accessibilityLabel="Назад"
-            onPress={() => router.back()}
-            className="flex-row items-center mb-2"
-            style={{ minHeight: 44 }}
-          >
-            <ChevronLeft size={20} color={colors.text} />
-            <Text className="text-text-base ml-1">Назад</Text>
-          </Pressable>
-          <Text className="text-2xl font-extrabold text-text-base mb-3">Создать заявку</Text>
-        </View>
-        <View className="flex-1 items-center justify-center">
-          <EmptyState
-            icon={Inbox}
-            title="Лимит активных заявок исчерпан (5/5)"
-            subtitle="Закройте одну активную заявку, чтобы создать новую"
-            actionLabel="Открыть мои заявки"
-            onAction={() => nav.any(ROUTES.tabsRequests)}
-          />
-        </View>
-      </SafeAreaView>
-    );
-  }
-
-  return (
-    <SafeAreaView className="flex-1 bg-surface2">
-      <View className="px-4 pt-4">
-        <Pressable
-          accessibilityRole="button"
-          accessibilityLabel="Назад"
-          onPress={() => router.back()}
-          className="flex-row items-center mb-2"
-          style={{ minHeight: 44 }}
-        >
-          <ChevronLeft size={20} color={colors.text} />
-          <Text className="text-text-base ml-1">Назад</Text>
-        </Pressable>
-        <Text className="text-2xl font-extrabold text-text-base mb-3">Создать заявку</Text>
-      </View>
-      <ScrollView
-        className="flex-1"
-        keyboardShouldPersistTaps="handled"
-        contentContainerStyle={{ paddingBottom: 32 }}
-      >
-        <View
-          style={{
-            width: "100%",
-            maxWidth: 720,
-            alignSelf: "center",
-            paddingHorizontal: 16,
-            paddingTop: 16,
-          }}
-        >
-          {!isAuthenticated && (
-            <View className="bg-accent-soft border border-accent rounded-xl p-4 mb-4">
-              <Text className="text-sm font-semibold text-accent mb-0.5">
-                Заполните заявку — войдёте в один шаг при отправке
-              </Text>
-              <Text className="text-sm text-accent">
-                Подтверждение по email кодом. Без паролей.
-              </Text>
-            </View>
-          )}
-
-          <View className="bg-white border border-border rounded-2xl px-4 pt-4 pb-4 mb-4">
-            <Text className="text-xs font-semibold text-text-mute uppercase tracking-wider mb-3">
-              Описание заявки
-            </Text>
-
-            <View className="mb-4">
-              <Text className="text-sm font-medium text-text-base mb-1.5">
-                Заголовок <Text className="text-danger">*</Text>
-              </Text>
-              <Input
-                placeholder="Кратко опишите суть проблемы"
-                value={title}
-                onChangeText={setTitle}
-                error={
-                  (submitted || title.length > 0) && !titleValid
-                    ? title.trim().length < 3
-                      ? "Минимум 3 символа"
-                      : "Максимум 100 символов"
-                    : undefined
-                }
-                maxLength={100}
-                editable={!atLimit && !submitting}
-              />
-              <Text
-                className={`text-xs text-right mt-1 ${
-                  title.length >= 100 ? "text-danger" : "text-text-dim"
-                }`}
-              >
-                {title.length}/100
-              </Text>
-            </View>
-
-            <CityFnsServicePicker
-              cities={cities}
-              fnsOffices={fnsOffices}
-              services={services}
-              selectedCity={selectedCity}
-              selectedFns={selectedFns}
-              selectedService={selectedService}
-              cityOpen={cityOpen}
-              fnsOpen={fnsOpen}
-              serviceOpen={serviceOpen}
-              loadingFns={loadingFns}
-              submitted={submitted}
-              disabled={atLimit || submitting}
-              onCitySelect={handleCitySelect}
-              onFnsSelect={handleFnsSelect}
-              onServiceSelect={handleServiceSelect}
-              onServiceClear={() => {
-                setSelectedServiceId(null);
-                setServiceOpen(false);
-              }}
-              onCityOpenChange={setCityOpen}
-              onFnsOpenChange={setFnsOpen}
-              onServiceOpenChange={setServiceOpen}
-            />
-
-            <View className="mb-4">
-              <Text className="text-sm font-medium text-text-base mb-1.5">
-                Описание <Text className="text-danger">*</Text>
-              </Text>
-              <Input
-                placeholder="Подробно опишите ситуацию: что произошло, какие документы получили, что требует инспекция, какая помощь нужна"
-                value={description}
-                onChangeText={setDescription}
-                multiline
-                error={
-                  (submitted || description.length > 0) && !descriptionValid
-                    ? description.trim().length < 10
-                      ? "Минимум 10 символов"
-                      : "Максимум 2000 символов"
-                    : undefined
-                }
-                maxLength={2000}
-                editable={!atLimit && !submitting}
-                containerStyle={{ minHeight: 120 }}
-              />
-              <Text
-                className={`text-xs text-right mt-1 ${
-                  description.length >= 2000 ? "text-danger" : "text-text-dim"
-                }`}
-              >
-                {description.length}/2000
-              </Text>
-            </View>
-          </View>
-
-          {submitError ? (
-            <View className="bg-danger-soft border border-danger rounded-xl p-3 mb-4">
-              <Text className="text-sm font-semibold text-danger mb-0.5">
-                Ошибка публикации
-              </Text>
-              <Text className="text-sm text-danger">{submitError}</Text>
-            </View>
-          ) : null}
-
-          {/* Inline OTP block — anon submit only */}
-          {!isAuthenticated && showOtpFlow && (
-            <InlineOtpFlow
-              onAuthenticated={submitRequestAuthed}
-              onCancel={() => setShowOtpFlow(false)}
-              parentSubmitting={submitting}
-              returnTo="/requests/create"
-            />
-          )}
-
-          {/* Primary submit button — hidden once inline OTP block is open. */}
-          {(isAuthenticated || !showOtpFlow) && (
-            <Button
-              label={
-                isAuthenticated
-                  ? "Опубликовать заявку"
-                  : "Отправить заявку"
-              }
-              onPress={handleSubmit}
-              disabled={submitting || atLimit}
-              loading={submitting}
-            />
-          )}
-        </View>
-      </ScrollView>
-    </SafeAreaView>
-  );
+/**
+ * #1518 — legacy /requests/create route. The canonical create-request
+ * screen lives at /requests/new. This file remains as a redirect so any
+ * deep-linked or bookmarked URL continues to work without a 404.
+ */
+export default function CreateRequestRedirect() {
+  return <Redirect href={"/requests/new" as never} />;
 }
